+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- force refreshing suggestions with `SuggestionsController.refresh`
+
 ## 5.1.0 - 2024-01-27
 ### Added
 - Returning null from `suggestionsCallback` hides the box